feat: adding abstraction layer for response conversions
response formats are now data, stored and live.  There also are int representations for each.  There
are 6 conversion functions, allowing any of the three formats to be converted to the other two.  In
addition, there are functions for int conversion, dealing with the issues on the driver of int
representation in js.

fix #17