fix(mocksessionfromqueryset): added console log for error of not matching in mockSessionFromQuerySet
refactored out error messages into some helper functions and added console logs of them.  At times,
it results in a redundant console log, but it seems to me to be worth it so that someone can
directly copy an int test into a unit test and run it with a mock driver, with the errors printing.

fix #21