feat: record transaction type
when we call a transaction, the output summary contains a `.transactionType` which is set to `READ`
or `WRITE`.

BREAKING CHANGE: If a generated session calls `readTransaction()` or `writeTransaction()`, the
result will have in it's `summary` a `transactionType` set to `READ` or `WRITE`.  So if you are
asserting deep equality between the output summary and expected summary, you will have to update the
expected summary accordingly.

fix #38